LLM-PHASE2/requirements.txt
streamlit
ollama
Pillow
openai
faster-whisper
whisper
feedparser>=6.0.0
beautifulsoup4
dateparser
requests

